bfportal/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
binaryornot==0.4.4
bleach==6.0.0
bpyutils==0.5.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
cryptography==39.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-ajax-selects @ git+https://github.com/crucialfelix/django-ajax-selects.git@9bf8df4b6c61d2f89a331d67337b77c09be94321
django-allauth==0.48.0
django-autocomplete-light==3.9.4
django-browser-reload==1.7.0
django-embed-video==1.4.8
django-extensions==3.2.1
django-filter==22.1
django-livereload-server==0.4
django-markdownx @ git+https://github.com/neutronX/django-markdownx.git@adcaa9e375a4691bbd16c4e27f8cdfb985366088
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-tailwind==3.5.0
django-treebeard==4.6.1
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
Faker==17.6.0
filelock==3.10.0
frozenlist==1.3.3
gunicorn==20.1.0
html5lib==1.1
identify==2.5.21
idna==3.4
isort==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
l18n==2021.3
loguricorn==0.1.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
nodeenv==1.7.0
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==9.4.0
pipreqs==0.4.11
pipupgrade==1.12.0
platformdirs==2.6.2
poyo==0.5.0
pre-commit==3.2.0
psycopg2==2.9.5
py-cord==2.4.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.15
virtualenv==20.16.3
wagtail==4.2.1
wagtail-autocomplete @ git+https://github.com/wagtail/wagtail-autocomplete.git@03f46a0c256989690d0a43fc21fe0f37f9ede765
wagtail-color-panel==1.3.1
wagtail-generic-chooser==0.5.1
wagtailsvg==0.0.37
webencodings==0.5.1
whitenoise==6.4.0
Willow==1.4.1
win32-setctime==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.9
xlwt==1.3.0
yarg==0.1.9
yarl==1.8.2
